Bug: switch from EntGDB to FileGDB, handle schema
Switching from an enterprise GDB to a file GDB doesn't handle schema (there are no schema's on File GDBs).  Added a test such that schema will be removed if the workspace type is set to File GDB.  Passing in an explicit dataset name should be unaffected, but passing in no name should just work (assuming no changes to the name, just stripping the schema/user).