Explicitly avoid/fail import of arcpyext.mapping
This module is only useful for arcpy on ArcGIS Desktop/Python 2.7.  On arcpy for ArcGIS Pro/Python 3.6, the relevant arcpy module is "mp", so arcpyext will have an accompanying module (arcpyext.mp) with similar functionality to arcpyext.mapping (as needed).

Currently, arcpyext.publishing._publishing has MXD-related functions that reach into arcpyext.mapping, so the whole module throws an exception when imported on Python 3.6.